+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED (YYYY-MM-DD)
 
+This is a minor release.
+
+### Added
+
+- `SingleTarget` now implements `Serialize`/`Deserialize`, following the `FromStr` implementation
+
 ## Version 0.3.0 (2025-08-27)
 
 This release focuses on supporting listing files contained in torrents. This is not implemented for magnet files, but is implemented for `TorrentFile` and will be implemented in [hightorrent_api](https://github.com/angrynode/hightorrent_api) for the QBittorrent backend.